requirements.txt
# Web framework & sessions
Flask==3.0.3
Flask-Session==0.6.0

# OpenAI SDK (newer 1.x API, supports openai.chat.completions.create)
openai>=1.30.0

# HTTP + SMS
requests>=2.32.0

# Env vars (recommended even if you currently hardcode keys)
python-dotenv>=1.0.1

#Discord bot
discord.py

# (Optional) Production server for deploys (not needed for local dev)
# gunicorn>=21.2.0
